refactor: refactor proj struct
