+++ parsers/parser_common.py
@@ -1,88 +1,17 @@
 import logging
 import traceback
-
+from datetime import datetime
+import os
 from selenium.common import NoSuchElementException, TimeoutException, ElementClickInterceptedException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 import time
-
-import parser_common
-
-
-def get_product_info_de(asin_list, country_name, amazon_url, postal_code, driver):
-    retries = 0
-    max_retries = 3
-    fail_flag = False
-    while retries <= max_retries:
-        try:
-            logging.info(f'打开 {country_name} 站点')
-            driver.get(amazon_url)
-
-            # 选择 Choose your location
-            # 刚打开网页后,这里会出现验证码校验,给多点时间,或许之后提供人工输入
-            logging.info(f'验证码校验处理-开始')
-            # 如果出现验证码校验,尝试点击 "Try different image" 按钮,可能会跳过验证码校验
-            captcha_jump = parser_common.captcha_jump(country_name, driver)
-            if captcha_jump is False:
-                fail_flag = True
-                break
-
-            logging.info(f'验证码校验处理-完成')
-
-            time.sleep(1)
-            logging.info(f'设置邮编-开始')
-
-            try:
-                choose_location_button = WebDriverWait(driver, 60).until(
-                    EC.element_to_be_clickable((By.ID, 'nav-global-location-popover-link'))
-                )
-                choose_location_button.click()
-            except (NoSuchElementException, TimeoutException, ElementClickInterceptedException) as e:
-                parser_common.except_screenshot(driver)
-                logging.error(f'出现的异常信息：{str(e)}')
-                logging.error(f'{country_name},重新点击设置邮编')
-                driver.execute_script("arguments[0].click;", choose_location_button)
-                pass
-
-            parser_common.except_screenshot(driver)
-            postal_code_input = WebDriverWait(driver, 10).until(
-                EC.presence_of_element_located((By.ID, 'GLUXZipUpdateInput'))
-            )
-            postal_code_input.clear()
-            postal_code_input.send_keys(postal_code)
-
-            # 点击 Apply
-            apply_button = WebDriverWait(driver, 10).until(
-                EC.element_to_be_clickable((By.CSS_SELECTOR, '#GLUXZipUpdate > span > input'))
-            )
-            apply_button.click()
-            logging.info(f'设置邮编-完成')
-
-            # 等待页面加载
-            time.sleep(1)
-            break
-
-        except Exception as e:
-            parser_common.except_screenshot(driver)
-            logging.error(f'当前名为{country_name}的 Excel 数据没能开始爬取数据')
-            logging.error("可能是出现 amazon 验证码校验提示,等待一段时间再启动程序吧")
-            logging.error(f'出现的异常信息：{str(e)}')
-            # 捕获异常，并输出具体位置和报错信息
-            logging.error(traceback.format_exc())
-            time.sleep(3)
-            retries += 1
-            if retries > 3:
-                return None
-            logging.info(f'amazon 反爬虫导致未能正常采集数据,第{retries}次重试...')
-
-    if fail_flag is True:
-        return None
-    else:
-        return get_common_product_data_set_multi_tabs_de(asin_list, country_name, amazon_url, driver)
+import random
+import secrets
 
 
-def get_common_product_data_set_multi_tabs_de(asin_list, country_name, amazon_url, driver):
+def get_common_product_data_set_multi_tabs(asin_list, country_name, amazon_url, driver):
     # 创建结果集合
     results = []
     logging.info(f'开始循环打开 {country_name} 商品')
@@ -101,8 +30,7 @@ def get_common_product_data_set_multi_tabs_de(asin_list, country_name, amazon_ur
             if i + j < len(asin_list):
                 driver.switch_to.window(handles[j])
 
-                single_result = get_common_product_data_set_single_de(asin_list[i + j], country_name, amazon_url,
-                                                                      driver)
+                single_result = get_common_product_data_set_single(asin_list[i + j], country_name, amazon_url, driver)
 
                 if single_result is not None:
                     results.append(single_result)
@@ -110,7 +38,7 @@ def get_common_product_data_set_multi_tabs_de(asin_list, country_name, amazon_ur
     return results
 
 
-def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver):
+def get_common_product_data_set_single(asin, country_name, amazon_url, driver):
     # 打开商品详情链接
     product_url = f'{amazon_url}dp/{asin}'
     driver.get(product_url)
@@ -125,11 +53,11 @@ def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver
     except TimeoutException:
         # 该 asin 商品的页面失效
         # 弹出验证码校验
-        parser_common.except_screenshot(driver)
+        except_screenshot(driver)
         logging.warning(f'{country_name}-{asin}, amazon 反爬,尝试绕过')
 
         # 尝试点击验证码绕过
-        captcha_asin = parser_common.captcha_jump(country_name, driver)
+        captcha_asin = captcha_jump(country_name, driver)
         if captcha_asin is False:
             logging.error(f'{country_name}-{asin}, 无法绕过, 如果频繁出现该信息, 请等待一段时间再运行')
             return None
@@ -146,11 +74,7 @@ def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver
     try:
         # 检查是否不可在当前区域售卖
         unavailable_element = driver.find_element(By.ID, "availability")
-
-        # if unavailable_element:
         if "Currently unavailable" in unavailable_element.text:
-            is_unavailable = True
-        elif "Derzeit nicht verfügbar" in unavailable_element.text:
             is_unavailable = True
 
     except NoSuchElementException:
@@ -160,13 +84,12 @@ def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver
 
     if is_unavailable:
         # 获取商品信息
-        title_element = parser_common.find_element_safe(country_name, asin, driver, By.ID, 'productTitle')
-        brand_element = parser_common.find_element_safe(country_name, asin, driver, By.ID, 'bylineInfo')
-        rating_element = parser_common.find_element_safe(country_name, asin, driver, By.ID, 'acrPopover')
-        rating_count_element = parser_common.find_element_safe(country_name, asin, driver, By.ID,
-                                                               'acrCustomerReviewText')
-        category_element = parser_common.find_element_safe(country_name, asin, driver, By.ID,
-                                                           'wayfinding-breadcrumbs_feature_div')
+        title_element = find_element_safe(country_name, asin, driver, By.ID, 'productTitle')
+        brand_element = find_element_safe(country_name, asin, driver, By.ID, 'bylineInfo')
+        rating_element = find_element_safe(country_name, asin, driver, By.ID, 'acrPopover')
+        rating_count_element = find_element_safe(country_name, asin, driver, By.ID, 'acrCustomerReviewText')
+        category_element = find_element_safe(country_name, asin, driver, By.ID,
+                                             'wayfinding-breadcrumbs_feature_div')
 
         title = "未找到标题"
         brand = "未找到品牌"
@@ -181,7 +104,8 @@ def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver
         if brand_element is not None:
             brand = brand_element.text.strip()
             # 'Brand: WindMax' -> 'WindMax'
-            brand = parser_common.amazon_brand_split(brand)
+            # brand = brand.replace('Brand:', '').strip()
+            brand = amazon_brand_split(brand)
             # Marke: BITOM
             # Marca: Compatibile
             # 品牌：Dalinch
@@ -223,6 +147,99 @@ def get_common_product_data_set_single_de(asin, country_name, amazon_url, driver
 
     logging.info(f'{country_name}-{asin} 完成数据采集')
     # 等待一段时间,防止访问频率过高被封禁或出现人机校验
-    time.sleep(1)
+    time.sleep(2)
 
     return result
+
+
+def find_element_safe(country_name, asin, driver, by, value):
+    try:
+        element = driver.find_element(by, value)
+        return element
+    except NoSuchElementException:
+        logging.info(f'{country_name}-{asin},页面不存在 {value} 元素')
+        return None
+
+
+def fake_operation(country_name, asin, driver, input_element):
+    search_element = find_element_safe(country_name, asin, driver, By.ID, input_element)
+
+    if search_element is not None:
+        # 模拟输入操作
+        search_element.clear()
+        search_element.send_keys(secrets.token_hex(10))
+        # 模拟随机滚动
+        scroll_position = random.randint(0, 900)
+        driver.execute_script("window.scrollTo(0, {});".format(scroll_position))
+        time.sleep(2)
+        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
+        driver.execute_script("window.scrollTo(0, 0);")
+
+
+def fake_operation_scroll(driver):
+    # 模拟随机滚动
+    scroll_position = random.randint(0, 900)
+    driver.execute_script("window.scrollTo(0, {});".format(scroll_position))
+    time.sleep(1)
+    driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
+    driver.execute_script("window.scrollTo(0, 0);")
+
+
+def captcha_jump(country_name, driver):
+    # 如果出现验证码校验,尝试点击 "Try different image" 按钮,可能会跳过验证码校验
+    try:
+        except_screenshot(driver)
+        captcha_switch_button = WebDriverWait(driver, 12).until(
+            EC.element_to_be_clickable((By.XPATH,
+                                        '/html/body/div/div[1]/div[3]/div/div/form/div[1]/div/div/div[2]/div/div[2]/a'))
+        )
+
+        captcha_switch_button.click()
+        except_screenshot(driver)
+        return True
+
+    except TimeoutException as e:
+        except_screenshot(driver)
+        logging.error(f'{country_name},未找到验证码元素')
+        logging.error(f'出现的异常信息：{str(e)}')
+        logging.error(traceback.format_exc())
+        return True
+
+    except (NoSuchElementException, ElementClickInterceptedException) as e:
+        except_screenshot(driver)
+        logging.error(f'{country_name},未找到 "Try different image" 按钮')
+        logging.error(f'{country_name},或者是 amazon 校验方式逻辑改变了')
+        logging.error(f'出现的异常信息：{str(e)}')
+        # 捕获异常，并输出具体位置和报错信息
+        logging.error(traceback.format_exc())
+        return False
+
+    except Exception as e:
+        except_screenshot(driver)
+        logging.error(f'{country_name},验证码校验失败')
+        logging.error(f'出现的异常信息：{str(e)}')
+        # 捕获异常，并输出具体位置和报错信息
+        logging.error(traceback.format_exc())
+        return False
+
+
+def except_screenshot(driver):
+    # 创建截图保存目录
+    screenshot_dir = os.path.join(os.getcwd(), '../log', 'img')
+    os.makedirs(screenshot_dir, exist_ok=True)
+    # 生成时间戳作为截图文件名
+    timestamp = datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
+    screenshot_path = os.path.join(screenshot_dir, f"{timestamp}.png")
+
+    # 获取当前浏览器的截图并保存
+    driver.save_screenshot(screenshot_path)
+
+
+def amazon_brand_split(text):
+    parts = text.split(":")
+
+    if len(parts) > 1:
+        return parts[-1].strip()
+
+    parts = text.split("：")
+    return parts[-1].strip()
