+++ parsers/parser_fr.py
@@ -1,16 +1,16 @@
 import logging
-import time
 import traceback
 
 from selenium.common import NoSuchElementException, TimeoutException, ElementClickInterceptedException
 from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+import time
 
-import parser_common
+import parsers.parser_common as parser_common
 
 
-def get_product_info_co_jp(asin_list, country_name, amazon_url, postal_code, driver):
+def get_product_info_fr(asin_list, country_name, amazon_url, postal_code, driver):
     retries = 0
     max_retries = 3
     fail_flag = False
@@ -46,19 +46,11 @@ def get_product_info_co_jp(asin_list, country_name, amazon_url, postal_code, dri
                 pass
 
             parser_common.except_screenshot(driver)
-            postal_code_split = postal_code.split("-")
-
             postal_code_input = WebDriverWait(driver, 10).until(
-                EC.presence_of_element_located((By.ID, 'GLUXZipUpdateInput_0'))
+                EC.presence_of_element_located((By.ID, 'GLUXZipUpdateInput'))
             )
             postal_code_input.clear()
-            postal_code_input.send_keys(postal_code_split[0])
-
-            postal_code_input = WebDriverWait(driver, 10).until(
-                EC.presence_of_element_located((By.ID, 'GLUXZipUpdateInput_1'))
-            )
-            postal_code_input.clear()
-            postal_code_input.send_keys(postal_code_split[1])
+            postal_code_input.send_keys(postal_code)
 
             # 点击 Apply
             apply_button = WebDriverWait(driver, 10).until(
@@ -87,10 +79,10 @@ def get_product_info_co_jp(asin_list, country_name, amazon_url, postal_code, dri
     if fail_flag is True:
         return None
     else:
-        return get_common_product_data_set_multi_tabs_co_jp(asin_list, country_name, amazon_url, driver)
+        return get_common_product_data_set_multi_tabs_fr(asin_list, country_name, amazon_url, driver)
 
 
-def get_common_product_data_set_multi_tabs_co_jp(asin_list, country_name, amazon_url, driver):
+def get_common_product_data_set_multi_tabs_fr(asin_list, country_name, amazon_url, driver):
     # 创建结果集合
     results = []
     logging.info(f'开始循环打开 {country_name} 商品')
@@ -109,8 +101,8 @@ def get_common_product_data_set_multi_tabs_co_jp(asin_list, country_name, amazon
             if i + j < len(asin_list):
                 driver.switch_to.window(handles[j])
 
-                single_result = get_common_product_data_set_single_co_jp(asin_list[i + j], country_name, amazon_url,
-                                                                         driver)
+                single_result = get_common_product_data_set_single_fr(asin_list[i + j], country_name, amazon_url,
+                                                                      driver)
 
                 if single_result is not None:
                     results.append(single_result)
@@ -118,7 +110,7 @@ def get_common_product_data_set_multi_tabs_co_jp(asin_list, country_name, amazon
     return results
 
 
-def get_common_product_data_set_single_co_jp(asin, country_name, amazon_url, driver):
+def get_common_product_data_set_single_fr(asin, country_name, amazon_url, driver):
     # 打开商品详情链接
     product_url = f'{amazon_url}dp/{asin}'
     driver.get(product_url)
@@ -156,14 +148,10 @@ def get_common_product_data_set_single_co_jp(asin, country_name, amazon_url, dri
         unavailable_element = driver.find_element(By.ID, "availability")
 
         # if unavailable_element:
-        if "目前无货" in unavailable_element.text:
+        if "Currently unavailable" in unavailable_element.text:
             is_unavailable = True
-        elif "Currently unavailable" in unavailable_element.text:
+        elif "Actuellement indisponible" in unavailable_element.text:
             is_unavailable = True
-        elif "In Stock" in unavailable_element.text:
-            is_unavailable = False
-        elif "现在有货" in unavailable_element.text:
-            is_unavailable = False
 
     except NoSuchElementException:
         # 未找到 "Currently unavailable." 元素,表示商品可在当前区域售卖
@@ -194,11 +182,10 @@ def get_common_product_data_set_single_co_jp(asin, country_name, amazon_url, dri
             brand = brand_element.text.strip()
             # 'Brand: WindMax' -> 'WindMax'
             brand = parser_common.amazon_brand_split(brand)
-            brand = remove_specific_substrings(brand)
             # Marke: BITOM
+            # Marca: Compatibile
             # 品牌：Dalinch
-            # 访问 Dikeshgu 品牌旗舰店
-            # 品牌：「Hello*Enjoy 専門店」
+            # Marque: Laurier
 
         if rating_element is not None:
             rating = rating_element.text.strip()
@@ -239,10 +226,3 @@ def get_common_product_data_set_single_co_jp(asin, country_name, amazon_url, dri
     time.sleep(1)
 
     return result
-
-
-def remove_specific_substrings(text):
-    substrings_to_remove = ["品牌旗舰店", "访问", "専門店", "「", "」"]
-    for substring in substrings_to_remove:
-        text = text.replace(substring, "")
-    return text.strip()
